http: Separate out net buffer code from formatting.
