+++ Common/Net/HTTPServer.cpp
@@ -35,10 +35,9 @@
 #include <cstdlib>
 
 #include "Common/Net/HTTPServer.h"
+#include "Common/Net/NetBuffer.h"
 #include "Common/Net/Sinks.h"
 #include "Common/File/FileDescriptor.h"
-
-#include "Common/Buffer.h"
 #include "Common/Log.h"
 
 void NewThreadExecutor::Run(std::function<void()> &&func) {
